Clean up and emergencies refactor
Got rid of admin controller
Removed refills
Removed marketing controller
Refactored emergencies to be like DHH
