--- app/assets/javascripts/refills/dropdown.js
@@ -1,12 +1,0 @@
-$(document).ready(function(){
-  $(".dropdown-button").click(function() {
-    $(".dropdown-menu").toggleClass("show-menu");
-    $(".dropdown-menu > li").click(function(){
-      $(".dropdown-menu").removeClass("show-menu");
-    });
-    $(".dropdown-menu.dropdown-select > li").click(function() {
-      $(".dropdown-button").html($(this).html());
-    });
-  });
-});
-
