+++ app/assets/stylesheets/admin.scss
@@ -18,7 +18,6 @@ $radio-empty-color: $light-gray;
 @import "base/base";
 @import "neat";
 @import "jquery.datetimepicker";
-@import "refills/*";
 @import "selectize";
 @import "selectize.default";
 @import "components/*";
