+++ app/assets/stylesheets/components/_pagination.scss
@@ -1,5 +1,5 @@
 .pagination {
-  ///////////////////////////////////////////////////////////////////////////////////
+  ///////////////////////////////////////////////////////////////////////
   $base-border-color: gainsboro !default;
   $base-border-radius: 3px !default;
   $base-spacing: 1.5em !default;
@@ -21,7 +21,7 @@
   a {
     text-decoration: none;
   }
-  //////////////////////////////////////////////////////////////////////////////////
+  ///////////////////////////////////////////////////////////////////////
 
   $pagination-border-color: $base-border-color;
   $pagination-border: 1px solid $pagination-border-color;
