--- app/assets/stylesheets/refills/_device.scss
@@ -1,101 +1,0 @@
-.device-background {
-  ///////////////////////////////////////////////////////////////////////////////////
-  $base-border-radius: 3px !default;
-  $base-accent-color: #477DCA !default;
-  $large-screen: em(860) !default;
-
-  p {
-    line-height: 1.5em;
-  }
-
-  h4 {
-    font-size: 1.5em;
-    margin: 0;
-  }
-  //////////////////////////////////////////////////////////////////////////////////
-
-  $device-padding-vertical: 5em;
-  $device-padding-horizontal: 0.4em;
-  $device-screen-width: 14em;
-  $device-screen-height: 25em;
-  $device-background: darken(gray, 10%);
-  $device-backside-background: darken($device-background, 10%);
-  $device-text-color: $base-font-color;
-  $device-background-top: #162C4C;
-  $device-background-bottom: #0A120D;
-  $gradient-angle: 10deg;
-  $device-image: 'tv_top.jpg';
-  $device-screen-image: 'triple_tv.jpg';
-
-  @include background(image-url($device-image), linear-gradient($gradient-angle, $device-background-bottom, $device-background-top), no-repeat $device-background-top scroll);
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center -263px;
-  min-height: 20em;
-  width: 100%;
-
-  .device-text {
-    color: $device-text-color;
-    padding: 2em;
-    text-align: center;
-
-    @include media($large-screen) {
-      padding: 4em;
-    }
-
-    h4 {
-      border-bottom: 1px solid transparentize($device-text-color, 0.7);
-      margin-bottom: 0.5em;
-      padding-bottom: 0.5em;
-    }
-
-    p {
-      margin-bottom: 1.5em;
-    }
-
-    button {
-      @include button(flat, $base-accent-color);
-      padding: 0.7em 1em;
-    }
-  }
-
-  .device {
-    display: none;
-    position: relative;
-
-    @include media($large-screen) {
-      @include transform(
-        perspective(800px)
-        translateX(0px)
-        translateY(-100px)
-        translateZ(50px)
-        rotateX(0deg)
-        rotateY(-20deg)
-        rotateZ(-0deg));
-      background: $device-background;
-      border-radius: 2em;
-      box-shadow:
-        1px 0px lighten($device-backside-background, 20%),
-        4px 0px lighten($device-backside-background, 2%),
-        7px 0px $device-backside-background,
-        10px 0px $device-backside-background,
-        13px 0px $device-backside-background;
-      display: block;
-      float: right;
-      margin-bottom: -12em;
-      margin-right: 5em;
-      padding: $device-padding-vertical 0;
-      width: $device-screen-width + (2 * $device-padding-horizontal);
-      z-index: 99999;
-
-      .screen {
-        @include size($device-screen-width $device-screen-height);
-        background-image: image-url($device-screen-image);
-        background-size: cover;
-        border-radius: 0.2em;
-        box-shadow: inset 0 1px 8px transparentize(black, 0.5);
-        margin: auto;
-      }
-    }
-  }
-}
