--- app/controllers/marketing_controller.rb
@@ -1,7 +1,0 @@
-class MarketingController < ApplicationController
-  skip_before_action :authenticate_user!
-  layout 'marketing'
-
-  def home
-  end
-end
