--- app/views/admin/index.html.erb
@@ -1,45 +1,0 @@
-<div class="admin-content">
-  <div class="content-container">
-    <div class="admin-site-title">
-      <h1>Digital Signage Administration</h1>
-      <p>The admin panel for Chapman's Digital Signs on Campus</p>
-      <br />
-    </div>
-
-    <div class="grid-items-lines">
-      <a id="manage-devices-link" href="<%= signs_path %>" class="grid-item">
-        <%= image_tag 'screen.png', alt: 'Signs' %>
-        <h1>Your Digital Signs</h1>
-        <p>Add, update, and delete digital signage devices.</p>
-      </a>
-
-      <a id="manage-slides-link" href="<%= slides_path %>" class="grid-item">
-        <%= image_tag 'film.png', alt: 'Slides' %>
-
-        <h1>Your Slides</h1>
-        <p>Add, update, and delete individual slides.</p>
-      </a>
-
-      <% if can? :broadcast_emergency, Sign %>
-      <a href="<%= admin_emergency_path%>" class="grid-item">
-        <%= image_tag 'feed.png', alt: 'Emergency' %>
-
-        <h1>Broadcast an Emergency</h1>
-        <p>Send an emergency notification to one, several, or all deivces.</p>
-      </a>
-      <% end %>
-
-      <% if can? :manage, User %>
-      <a href="<%= users_path %>" class="grid-item">
-        <%= image_tag 'users.png', alt: 'Users' %>
-
-        <h1>Manage User Permissions</h1>
-        <p>Allow users to manage devices and slides.</p>
-      </a>
-      <% end %>
-
-      <div class="right-cover"></div>
-      <div class="bottom-cover"></div>
-    </div>
-  </div>
-</div>
