--- app/views/admin/sessions.html.erb
@@ -1,1 +1,0 @@
-<%= session.to_hash %>
