--- app/views/marketing/home.html.erb
@@ -1,3 +1,0 @@
-<h1>Chapman University Digital Signage</h1>
-<p>A website...that goes on a TV.</p>
-<%= link_to "Get Started", signs_path %>
