--- app/views/refills/_breadcrumbs.html.erb
@@ -1,7 +1,0 @@
-<div class="breadcrumb">
-  <a href="javascript:void(0)">One</a>
-  <a href="javascript:void(0)">Two</a>
-  <a href="javascript:void(0)">Three</a>
-  <a href="javascript:void(0)">Four</a>
-  <a href="javascript:void(0)">Five</a>
-</div>
