--- app/views/refills/_device.html.erb
@@ -1,10 +1,0 @@
-<div class="device-background">
-  <div class="device-text">
-    <h4>Text about App or Device</h4>
-    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Dignissimos, recusandae? Deserunt repellendus.</p>
-    <button>Learn More</button>
-  </div>
-  <div class="device">
-    <div class="screen"></div>
-  </div>
-</div>
