--- app/views/refills/_dropdown.html.erb
@@ -1,12 +1,0 @@
-<div class="dropdown">
-  <div class="dropdown-container">
-    <p class="dropdown-description">Description</p>
-    <p class="dropdown-button">Click to Select</p>
-    <ul class="dropdown-menu dropdown-select">
-      <li>Dropdown Menu Item</li>
-      <li>Another Item</li>
-      <li>Item Three</li>
-      <li>The Last One</li>
-    </ul>
-  </div>
-</div>
