--- app/views/refills/_flashes.html.erb
@@ -1,15 +1,0 @@
-<div class="flash-success">
-  <span>This is a success message <a href="#">with a link</a></span>
-</div>
-
-<div class="flash-error">
-  <span>This is an error message <a href="#">with a link</a></span>
-</div>
-
-<div class="flash-notice">
-  <span>This is an notice message <a href="#">with a link</a></span>
-</div>
-
-<div class="flash-alert">
-  <span>This is an alert message <a href="#">with a link</a></span>
-</div>
