--- app/views/refills/_pagination.html.erb
@@ -1,17 +1,0 @@
-<div class="pagination">
-  <ul>
-    <li class="page-prev"><a href="javascript:void(0)">Prev</a></li>
-    <li>
-      <ul>
-        <li><a href="javascript:void(0)">1</a></li>
-        <li><a href="javascript:void(0)">2</a></li>
-        <li><a href="javascript:void(0)">3</a></li>
-        <li><a href="javascript:void(0)">4</a></li>
-        <li><a href="javascript:void(0)">5</a></li>
-        <li><a href="javascript:void(0)">6</a></li>
-        <li><a href="javascript:void(0)">7</a></li>
-      </ul>
-    </li>
-    <li class="page-next"><a href="javascript:void(0)">Next</a></li>
-  </ul>
-</div>
