--- app/views/refills/_search_bar.html.erb
@@ -1,6 +1,0 @@
-<form class="search-bar" role="search">
-  <input type="search" placeholder="Enter Search" />
-  <button type="submit">
-    <%= image_tag 'search-icon.png', alt: 'Search Icon' %>
-  </button>
-</form>
