--- app/views/refills/_switch.html.erb
@@ -1,4 +1,0 @@
-<label class="label-switch">
-  <input type="checkbox" />
-  <div class="checkbox"></div>
-</label>
