--- app/views/refills/_tooltip.html.erb
@@ -1,6 +1,0 @@
-<div class="tooltip-item">
-  Hover for Tooltip
-  <div class="tooltip">
-    <p>Lorem ipsum doldae oluptate aperiam unde voluptates quas.</p>
-  </div>
-</div>
