--- app/views/refills/_video.html.erb
@@ -1,5 +1,0 @@
-<div class="video">
-  <div class="video-wrapper">
-      <iframe src="http://www.youtube.com/embed/8ItNE_DX6Cc?showinfo=0&iv_load_policy=3&controls=0" frameborder="0" allowfullscreen></iframe>
-  </div>
-</div>
