--- test/controllers/marketing_controller_test.rb
@@ -1,10 +1,0 @@
-require 'test_helper'
-
-class MarketingControllerTest < ActionController::TestCase
-
-  test "should get home" do
-    get :home
-    assert_response :success
-  end
-
-end
